Move params.cpp and params.hpp to new locations
