+++ GPTifier/CMakeLists.txt
@@ -90,11 +90,11 @@ add_executable(
   "src/interface/command_run.cpp"
   "src/interface/command_short.cpp"
   "src/interface/command_test.cpp"
+  "src/interface/params.cpp"
   "src/main.cpp"
   "src/networking/api_openai_admin.cpp"
   "src/networking/api_openai_user.cpp"
   "src/networking/curl_base.cpp"
-  "src/params.cpp"
   "src/selectors.cpp"
   "src/serialization/chat_completions.cpp"
   "src/serialization/costs.cpp"
