+++ GPTifier/src/interface/command_costs.cpp
@@ -1,7 +1,7 @@
 #include "interface/command_costs.hpp"
 
 #include "help_messages.hpp"
-#include "params.hpp"
+#include "interface/params.hpp"
 #include "serialization/costs.hpp"
 #include "utils.hpp"
