+++ GPTifier/src/interface/command_run.cpp
@@ -2,7 +2,7 @@
 
 #include "datadir.hpp"
 #include "help_messages.hpp"
-#include "params.hpp"
+#include "interface/params.hpp"
 #include "selectors.hpp"
 #include "serialization/chat_completions.hpp"
 #include "utils.hpp"
