+++ GPTifier/src/interface/params.cpp
@@ -1,4 +1,4 @@
-#include "params.hpp"
+#include "interface/params.hpp"
 
 #include <fmt/core.h>
 #include <stdexcept>
